+++ d2c_robot_libs/src/serving_command.cpp
@@ -7,26 +7,26 @@ ServingCommand::~ServingCommand()
 {}
 
 
-std::vector<double> ServingCommand::ReturnTargetPosition()
+std::vector<float> ServingCommand::ReturnTargetPosition()
 {
-    std::vector<double> target_joint_position_ = {1.52, 0.0, 0.0, 0.0};
+    std::vector<float> target_joint_position_ = {1.52, 0.0, 0.0, 0.0};
     //std::vector<double> object_position_;
     //object_position_ = SetTargetPosition();
     //target_joint_position_ = InverseKinematics(object_position_);
 
     return target_joint_position_;
 }
-std::vector<double> ServingCommand::InverseKinematics(std::vector<double> object_position)
+std::vector<float> ServingCommand::InverseKinematics(std::vector<float> object_position)
 {
-    std::vector<double> target_joint_position;
+    std::vector<float> target_joint_position;
 
     //TODO : calculation 
 
     return target_joint_position;
 }
-std::vector<double> ServingCommand::SetTargetPosition()
+std::vector<float> ServingCommand::SetTargetPosition()
 {
-    std::vector<double> object_position;
+    std::vector<float> object_position;
 
     //TODO : get the information of object, and save into object_position variable
     //TODO :
